Fixed feature model editor bug when opening a .velvet model.

+++ plugins/de.ovgu.featureide.fm.core/src/de/ovgu/featureide/fm/core/io/velvet/VelvetFeatureModelReader.java
@@ -457,8 +457,10 @@ public class VelvetFeatureModelReader extends AbstractFeatureModelReader {
 		extFeatureModel.getLayout().showHiddenFeatures(true);
 		extFeatureModel.getLayout().verticalLayout(false);
 		
-		modelMarkerHandler = new ModelMarkerHandler(getProject().getFile(getFile().getName()));
-		modelMarkerHandler.deleteAllModelMarkers();
+		if (getProject() != null) {
+			modelMarkerHandler = new ModelMarkerHandler(getProject().getFile(getFile().getName()));
+			modelMarkerHandler.deleteAllModelMarkers();
+		}
 
 		extFeatureModelName = null;
 		cinterface = false;
@@ -1000,14 +1002,18 @@ public class VelvetFeatureModelReader extends AbstractFeatureModelReader {
 	}
 	
 	private void reportSyntaxError(Tree curNode) {
-		String message = "Illegal statement \""+ curNode.getText() + "\"";
-		modelMarkerHandler.createModelMarker(message, org.eclipse.core.resources.IMarker.SEVERITY_ERROR, curNode.getLine());
+		final String message = "Illegal statement \""+ curNode.getText() + "\"";
+		if (modelMarkerHandler != null) {
+			modelMarkerHandler.createModelMarker(message, org.eclipse.core.resources.IMarker.SEVERITY_ERROR, curNode.getLine());
+		}
 		FMCorePlugin.getDefault().logError(new UnsupportedModelException(
 			message + " at line " + curNode.getLine() + ((featureModelFile != null)?" in file " + featureModelFile.getName():""), curNode.getLine()));
 	}
 	
 	private void reportWarning(Tree curNode, String message) {
-		modelMarkerHandler.createModelMarker(message, org.eclipse.core.resources.IMarker.SEVERITY_WARNING, curNode.getLine());
+		if (modelMarkerHandler != null) {
+			modelMarkerHandler.createModelMarker(message, org.eclipse.core.resources.IMarker.SEVERITY_WARNING, curNode.getLine());
+		}
 		FMCorePlugin.getDefault().logWarning(message + " (at line "+ curNode.getLine() + ((featureModelFile != null)?" in file " + featureModelFile.getName():"") + ": \"" + curNode.getText() + "\")");		
 	}
 }
